+++ app/packs/funds/capital_commitments/services/import_fund_docs.rb
@@ -60,6 +60,9 @@ class ImportFundDocs < ImportUtil
                            name: user_data["Document Name"].strip, tag_list: user_data["Tags"]&.strip,
                            user_id: import_upload.user_id, send_email:)
 
+        # Save the document
+        raise "#{file_name} does not exist. Check for missing file or extension" unless File.exist?(file_name)
+
         doc.file = File.open(file_name, "rb")
         doc.save ? [true, "Success"] : [false, doc.errors.full_messages.join(", ")]
       end
