+++ app/packs/secondaries/offers/services/import_offer_docs.rb
@@ -57,6 +57,8 @@ class ImportOfferDocs < ImportUtil
                            name: user_data["Document Name"].strip, tag_list: user_data["Tags"]&.strip,
                            user_id: import_upload.user_id, send_email:)
 
+        raise "#{file_name} does not exist. Check for missing file or extension" unless File.exist?(file_name)
+
         doc.file = File.open(file_name, "rb")
         doc.save ? [true, "Success"] : [false, doc.errors.full_messages.join(", ")]
       end
